gpmc init: Clear gpmc_config1 register for all CS
Reset the CONFIG0 register, especially device_type field.
This is required to make decesions on interfaced flash devices
later in the boot process.

Kernel reads the DEVICETYPE field in detection of NOR or NAND flash
and the reset value of this field is 0, which is NOR. This creates
problem when you have NOR flash as a data flash connected to
different CS.